move environment map prefilter command recording code into Renderer
