+++ Source/Renderer/Renderer.h
@@ -52,6 +52,8 @@ struct FPostProcessParameters;
 struct ID3D12RootSignature;
 struct ID3D12PipelineState;
 struct TextureCreateDesc;
+struct FEnvironmentMapRenderingResources;
+struct Mesh;
 
 //========================================================================================================================================================================================================---
 //
@@ -201,7 +203,7 @@ public:
 	// ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 	// Render (Public)
 	// ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
+	void PreFilterEnvironmentMap(ID3D12GraphicsCommandList* pCmd, FEnvironmentMapRenderingResources& env, const Mesh& CubeMesh, HWND hwnd);
 
 
 	// ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -326,7 +328,7 @@ private:
 	// ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 	// Render (Private)
 	// ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-	
+	void ComputeBRDFIntegrationLUT(ID3D12GraphicsCommandList* pCmd, SRV_ID& outSRV_ID);
 
 //
 // STATIC PUBLIC DATA/INTERFACE
